Fixed loading the page when the background images record did not exist.
Fixed authentication handler -- the passport.authenticate function does not check if the user is logged in. Must use your own little middleware piece for that.

Added a logout route.

Added information to the main page for logged in users, including a link to the dashboard and a link to logout.